+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial creation of the provider against v0.5.7 of the OpsGenie Terraform Provider
 
 ---